% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringFunc.R
\name{title_case}
\alias{title_case}
\title{Title case a string}
\usage{
title_case(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
Character vector with first letter capitalized
}
\description{
Title case a string
}
